pyproject.toml
[tool.poetry]
name = "oil-price-notify"
version = "0.1.0"
description = ""
authors = ["DoubleChuang <ethan9141@gmail.com>"]
readme = "README.md"
packages = [{include = "oil_price_notify"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
lxml = "^4.9.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
